setting up basic CRUD function to mongoDB

+++ backend/controllers/orderController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require("express-async-handler");
 
 const Order = require("../models/orderModel");
+const User = require("../models/userModel");
 
 // @desc    Get all orders
 // @route   GET /api/orders
@@ -10,7 +11,7 @@ const getOrders = asyncHandler(async (req, res, next) => {
 		res.status(400);
 		next(new Error("Bad request"));
 	} else {
-		const orders = await Order.find();
+		const orders = await Order.find({ user: req.user._id });
 		res.status(200).send(orders);
 	}
 });
@@ -27,7 +28,7 @@ const getOrderById = asyncHandler(async (req, res) => {
 // @access  Private
 const createOrder = asyncHandler(async (req, res) => {
 	const { text } = req.body;
-	const order = await Order.create({ text });
+	const order = await Order.create({ text, user: req.user._id });
 	res.status(200).send(order);
 });
 
@@ -36,15 +37,29 @@ const createOrder = asyncHandler(async (req, res) => {
 // @access  Private/Admin
 const updateOrder = asyncHandler(async (req, res) => {
 	const order = await Order.findById(req.params.id);
-	const { text } = req.body;
-	if (order) {
-		order.text = text;
-		const updatedOrder = await order.save();
-		res.status(200).send(updatedOrder);
-	} else {
+	if (!order) {
 		res.status(404);
 		throw new Error("Order not found");
 	}
+
+	const user = await User.findById(req.user._id);
+
+	// Check if user exists
+	if (!user) {
+		res.status(404);
+		throw new Error("User not found");
+	}
+
+	// Check if user is the owner of the order
+	if (order.user.toString() !== req.user._id.toString()) {
+		res.status(401);
+		throw new Error("User not authorized");
+	}
+
+	// Update order
+	const updateOrder = await Order.findByIdAndUpdate(req.params.id, req.body, {
+		new: true,
+	});
 });
 
 // @desc    Delete order
@@ -52,12 +67,30 @@ const updateOrder = asyncHandler(async (req, res) => {
 // @access  Private/Admin
 const deleteOrder = asyncHandler(async (req, res) => {
 	const order = await Order.findById(req.params.id);
+
+	// Check if order exists
 	if (!order) {
 		res.status(404);
 		throw new Error("Order not found");
 	}
+
+	const user = await User.findById(req.user._id);
+
+	// Check if user exists
+	if (!user) {
+		res.status(404);
+		throw new Error("User not found");
+	}
+
+	// Check if user is the owner of the order
+	if (order.user.toString() !== req.user._id.toString()) {
+		res.status(401);
+		throw new Error("User not authorized");
+	}
+
+	// Delete order
 	await order.deleteOne();
-	res.status(200).json({ message: `Order ${req.params.id} removed` });
+	res.status(200).send("Order deleted");
 });
 
 module.exports = {
